# BE/docs/swagger.yaml
basePath: /api
...
info:
  contact:
    email: support@seapedia.com
    name: API Support
  description: API Documentation for SEAPEDIA platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SEAPEDIA API
  version: "1.0"
paths:
  /api/addresses:
    get:
      description: Get a list of addresses for the current user
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get User Addresses
      tags:
      - Addresses
    post:
      consumes:
      - application/json
      description: Add a new address for the current user
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create Address
      tags:
      - Addresses
  /api/addresses/{id}:
    delete:
      description: Delete an address
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete Address
      tags:
      - Addresses
    put:
      consumes:
      - application/json
      description: Update an existing address
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update Address
      tags:
      - Addresses
...
  /api/buyer/orders:
    get:
      description: Get a list of orders for the current buyer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List Buyer Orders
      tags:
      - Buyer Orders
  /api/buyer/orders/{id}:
    get:
      description: Get details of a specific order for the current buyer
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get Buyer Order Detail
      tags:
      - Buyer Orders
...
  /api/checkout:
    post:
      consumes:
      - application/json
      description: Create the order and process checkout
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Process Checkout
      tags:
      - Checkout
  /api/checkout/_preview:
    post:
      consumes:
      - application/json
      description: Calculate total, shipping, and discounts before checking out
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Preview Checkout
      tags:
      - Checkout
...
  /api/seller/promos:
    get:
      description: Get a list of promos for the current seller's store
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List Store Promos
      tags:
      - Seller Promos
    post:
      consumes:
      - application/json
      description: Create a new promo for the store
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a Promo
      tags:
      - Seller Promos
  /api/seller/promos/{id}:
    get:
      description: Get details of a specific promo
      parameters:
      - description: Promo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get Promo Detail
      tags:
      - Seller Promos
...
  /api/seller/vouchers:
    get:
      description: Get a list of vouchers for the current seller's store
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List Store Vouchers
      tags:
      - Seller Vouchers
    post:
      consumes:
      - application/json
      description: Create a new voucher for the store
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a Voucher
      tags:
      - Seller Vouchers
  /api/seller/vouchers/{id}:
    get:
      description: Get details of a specific voucher
      parameters:
      - description: Voucher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get Voucher Detail
      tags:
      - Seller Vouchers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"